# central-sso/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SSO Authentication API",
        "description": "Single Sign-On API for multi-tenant authentication",
        "contact": {
            "name": "SSO Admin",
            "email": "admin@sso.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/api/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Get all roles",
                "operationId": "ca86f0a206144dfbd695b749b4b28f9d",
                "parameters": [
                    {
                        "name": "include_permissions",
                        "in": "query",
                        "description": "Include permissions in response",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    },
                    {
                        "name": "include_system",
                        "in": "query",
                        "description": "Include system roles",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/RoleResponseDTO"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Create a new role",
                "operationId": "341b8b162b61267c7bd187500aea5bc4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRoleRequestDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Role created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Role created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/RoleResponseDTO"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/roles/{id}": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Get a specific role",
                "operationId": "15c404caf238f1861c81f68205d328e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Role ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/RoleResponseDTO"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Role not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Roles"
                ],
                "summary": "Update a role",
                "operationId": "c73ab816ae8fb7c34a995d37ba577082",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Role ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRoleRequestDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Role updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Role updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/RoleResponseDTO"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Cannot update system role"
                    },
                    "404": {
                        "description": "Role not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Roles"
                ],
                "summary": "Delete a role",
                "operationId": "ffa6ef115d86b00cc6c54a6210019f9c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Role ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Role deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Cannot delete system role"
                    },
                    "404": {
                        "description": "Role not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateRoleRequestDTO": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "Role name",
                        "type": "string",
                        "example": "Manager"
                    },
                    "slug": {
                        "description": "Role slug (auto-generated if not provided)",
                        "type": "string",
                        "example": "manager"
                    },
                    "description": {
                        "description": "Role description",
                        "type": "string",
                        "example": "Can manage users and view reports"
                    },
                    "permissions": {
                        "description": "Array of permission slugs to assign to this role",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "users.view",
                            "users.create",
                            "reports.view"
                        ]
                    }
                },
                "type": "object"
            },
...
            "PermissionResponseDTO": {
                "properties": {
                    "id": {
                        "description": "Permission ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Permission name",
                        "type": "string",
                        "example": "View Users"
                    },
                    "slug": {
                        "description": "Permission slug",
                        "type": "string",
                        "example": "users.view"
                    },
                    "description": {
                        "description": "Permission description",
                        "type": "string",
                        "example": "Allows viewing user list and details"
                    },
                    "category": {
                        "description": "Permission category",
                        "type": "string",
                        "example": "users"
                    },
                    "is_system": {
                        "description": "Whether the permission is a system permission",
                        "type": "boolean",
                        "example": false
                    },
                    "roles_count": {
                        "description": "Number of roles with this permission",
                        "type": "integer",
                        "example": 3
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-16T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-16T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "RoleResponseDTO": {
                "properties": {
                    "id": {
                        "description": "Role ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Role name",
                        "type": "string",
                        "example": "Manager"
                    },
                    "slug": {
                        "description": "Role slug",
                        "type": "string",
                        "example": "manager"
                    },
                    "description": {
                        "description": "Role description",
                        "type": "string",
                        "example": "Can manage users and view reports"
                    },
                    "is_system": {
                        "description": "Whether the role is a system role",
                        "type": "boolean",
                        "example": false
                    },
                    "permissions": {
                        "description": "List of permissions assigned to this role",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PermissionResponseDTO"
                        }
                    },
                    "users_count": {
                        "description": "Number of users with this role",
                        "type": "integer",
                        "example": 5
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-16T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-16T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}